Rename osx to bsd
