+++ setup.py
@@ -381,7 +381,7 @@ setup(name=APPNAME,
                 'allmydata.util',
                 'allmydata.web',
                 'allmydata.windows',
-                'allmydata.osx',
+                'allmydata.bsd',
                 'buildtest'],
       classifiers=trove_classifiers,
       test_suite="allmydata.test",
